Move suites logic into ./lib/suites.js module.
