+++ lib/index.js
@@ -11,6 +11,7 @@ import {
   addRoutes as addContextStoreRoutes
 } from '@bedrock/service-context-store';
 import {addRoutes} from './http.js';
+import {getSuiteParams} from './suites.js';
 import {initializeServiceAgent} from '@bedrock/service-agent';
 import {klona} from 'klona';
 
@@ -93,7 +94,7 @@ async function validateConfigFn({config} = {}) {
 
     // ensure suite parameters can be retrieved for the configured
     // `issueOptions` and `statusListOptions`...
-    issuer.getSuiteParams({config, suiteName: issueOptions.suiteName});
+    getSuiteParams({config, suiteName: issueOptions.suiteName});
     for(const {type, statusPurpose, suiteName} of statusListOptions) {
       // `statusPurpose` MUST be `revocation` if using type `RevocationList2020`
       if(type === 'RevocationList2020' && statusPurpose !== 'revocation') {
@@ -102,7 +103,7 @@ async function validateConfigFn({config} = {}) {
           '"revocation".');
       }
 
-      issuer.getSuiteParams({config, suiteName});
+      getSuiteParams({config, suiteName});
     }
   } catch(error) {
     return {valid: false, error};
