+++ lib/issuer.js
@@ -8,12 +8,10 @@ import {AsymmetricKey, KmsClient} from '@digitalbazaar/webkms-client';
 import {documentStores, serviceAgents} from '@bedrock/service-agent';
 import assert from 'assert-plus';
 import {createDocumentLoader} from './documentLoader.js';
-import {createEddsa2022Suite} from './suites.js';
 import {CredentialStatusWriter} from './CredentialStatusWriter.js';
 import {decodeList} from '@digitalbazaar/vc-status-list';
 import {didIo} from '@bedrock/did-io';
-import {Ed25519Signature2018} from '@digitalbazaar/ed25519-signature-2018';
-import {Ed25519Signature2020} from '@digitalbazaar/ed25519-signature-2020';
+import {getSuiteParams} from './suites.js';
 import {httpsAgent} from '@bedrock/https-agent';
 import {logger} from './logger.js';
 import {constants as rlConstants} from '@bedrock/vc-revocation-list-context';
@@ -22,41 +20,9 @@ import {constants as slConstants} from '@bedrock/vc-status-list-context';
 const {util: {BedrockError}} = bedrock;
 
 const serviceType = 'vc-issuer';
-const SUPPORTED_SUITES = new Map([
-  ['Ed25519Signature2020', {
-    keyType: 'ed25519', createSuite:
-      ({signer}) => new Ed25519Signature2020({signer})
-  }],
-  ['Ed25519Signature2018', {
-    keyType: 'ed25519', createSuite:
-      ({signer}) => new Ed25519Signature2018({signer})
-  }],
-  ['eddsa-2022', {
-    keyType: 'ed25519', createSuite: createEddsa2022Suite
-  }]
-]);
 
 // exported for testing purposes
 export const _CredentialStatusWriter = CredentialStatusWriter;
-
-export function getSuiteParams({config, suiteName}) {
-  // ensure suite is supported
-  const suiteInfo = SUPPORTED_SUITES.get(suiteName);
-  if(!suiteInfo) {
-    throw new Error(`Unsupported suite "${suiteName}".`);
-  }
-
-  // get zcap to use to invoke assertion method key
-  const {keyType, createSuite} = suiteInfo;
-  const referenceId = `assertionMethod:${keyType}`;
-  const zcap = config.zcaps[referenceId];
-  if(!zcap) {
-    throw new Error(
-      `No capability available to sign using suite "${suiteName}".`);
-  }
-
-  return {zcap, createSuite, referenceId};
-}
 
 export async function issue({credential, config} = {}) {
   assert.object(credential, 'credential');
